+++ api/lib/block-listeners/cosmos-node-subscription.js
@@ -100,6 +100,8 @@ class CosmosNodeSubscription {
       console.error('Failed to fetch block', error)
       Sentry.captureException(error)
     }
+    // overwrite chain_id with the network's one, making sure it is correct
+    this.store.network.chain_id = block.chainId
     if (block && this.height !== block.height) {
       // apparently the cosmos db takes a while to serve the content after a block has been updated
       // if we don't do this, we run into errors as the data is not yet available