+++ api/lib/routes/transaction.js
@@ -1,6 +1,6 @@
 var express = require('express')
 var router = express.Router()
-var { estimate, broadcast } = require('./../controller/transaction')
+var { broadcast } = require('./../controller/transaction')
 
 router.use(function timeLog(req, res, next) {
   req.txRequest = req.body && req.body.payload
@@ -10,11 +10,6 @@ router.use(function timeLog(req, res, next) {
     res.json({ error: 'No Request Found' })
   }
   next()
-})
-
-router.use('/estimate', async function (req, res) {
-  const response = await estimate(req.txRequest)
-  res.json(response)
 })
 
 router.use('/broadcast', async function (req, res) {